requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytesseract
pandas
pyyaml
python-docx
PyMuPDF
transformers
torch
spacy
sqlalchemy
alembic
openpyxl
python-dotenv
celery[redis]
redis
